% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAGUtils.R
\name{estimatePAG}
\alias{estimatePAG}
\title{Estimates the PAG using the FCI algorithm from the \code{pcalg} R package
It supports datasets with all variables binary or Gaussian.}
\usage{
estimatePAG(dat, Vnames, isbinary, alpha)
}
\arguments{
\item{dat}{a data-frame with the observed data.}

\item{Vnames}{names os the variables in the dataset, usually the column names of dat.}

\item{isbinary}{Boolean if TRUE, then it is expected that all variables are binary. If FALSE,
the it is expected that are variables follow a Gaussian distribution.}

\item{alpha}{the required significance level for the conditional independence tests.}
}
\value{
a \code{pcalg} object resulting from the FCI algorithm.
}
\description{
Estimates the PAG using the FCI algorithm from the \code{pcalg} R package
It supports datasets with all variables binary or Gaussian.
}
