% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAGUtils.R
\name{getBucket}
\alias{getBucket}
\title{Gets the bucket of the node 'node'}
\usage{
getBucket(amat, node)
}
\arguments{
\item{amat}{the \code{pcalg}-compatible adjacency matrix of the PAG}

\item{node}{name of the node}
}
\value{
the bucket of the node 'node' in the PAG represented by its adjacency matrix 'amat'
}
\description{
Gets the bucket of the node 'node'
}
