% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAGUtils.R
\name{getBucketList}
\alias{getBucketList}
\title{Gets the list of all buckets in the PAG}
\usage{
getBucketList(amat)
}
\arguments{
\item{amat}{the \code{pcalg}-compatible adjacency matrix of the PAG}
}
\value{
a list of all buckets in the PAG represented by its adjacency matrix 'amat'
}
\description{
Gets the list of all buckets in the PAG
}
