% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAGUtils.R
\name{getCCNodes}
\alias{getCCNodes}
\title{Gets all nodes that are connected to the set 'nodes' by
circle-circle edges in the PAG represented by 'amat'}
\usage{
getCCNodes(amat, nodes, excludenodes)
}
\arguments{
\item{amat}{the \code{pcalg}-compatible adjacency matrix of the PAG}

\item{nodes}{names of a set of nodes}

\item{excludenodes}{names of the nodes to be excluded.}
}
\value{
nodes that are connected to the set 'nodes' by
circle-circle edges
}
\description{
Gets all nodes that are connected to the set 'nodes' by
circle-circle edges in the PAG represented by 'amat'
}
