% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAGUtils.R
\name{getDagittyPAG}
\alias{getDagittyPAG}
\title{Converts a \code{pcalg} PAG into a \code{dagitty} PAG}
\usage{
getDagittyPAG(amat)
}
\arguments{
\item{amat}{\code{pcalg}-compatible adjacency matrix}
}
\value{
a \code{dagitty} object for the PAG
}
\description{
Converts a \code{pcalg} PAG into a \code{dagitty} PAG
}
