% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAGUtils.R
\name{getInvPossChildrenA}
\alias{getInvPossChildrenA}
\title{Gets the "invisible" children of A, i.e.,
the set \{ V :  A o-> V or A -> V with an invisible edge\}}
\usage{
getInvPossChildrenA(amat_v, amat, nodeA)
}
\arguments{
\item{amat_v}{TODO}

\item{amat}{the \code{pcalg}-compatible adjacency matrix of the PAG}

\item{nodeA}{name of the variable A}
}
\value{
names of all "invisible" possible children of A
}
\description{
Gets the "invisible" children of A, i.e.,
the set \{ V :  A o-> V or A -> V with an invisible edge\}
}
