% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAGUtils.R
\name{getPCComponentA}
\alias{getPCComponentA}
\title{Gets the possible C-Component of A, i.e.,}
\usage{
getPCComponentA(amat_v, amat, nodesA)
}
\arguments{
\item{amat_v}{TODO}

\item{amat}{the \code{pcalg}-compatible adjacency matrix of the PAG}

\item{nodesA}{name of the variables in the set A}
}
\value{
names of all "invisible" possible children of A
}
\description{
Gets the possible C-Component of A, i.e.,
}
