% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAGUtils.R
\name{getPossAncestors}
\alias{getPossAncestors}
\title{Gets all possible ancestors of Y in the PAG represented by the adjacency matrix amat}
\usage{
getPossAncestors(amat, ynames)
}
\arguments{
\item{amat}{the \code{pcalg}-compatible adjacency matrix of the PAG}

\item{ynames}{names of the variables in set Y}
}
\value{
names of all possible ancestors of Y
}
\description{
Gets all possible ancestors of Y in the PAG represented by the adjacency matrix amat
}
