% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAGUtils.R
\name{getPossDescendants}
\alias{getPossDescendants}
\title{Gets all possible descendants of Y in the PAG represented by the adjacency matrix amat}
\usage{
getPossDescendants(amat, ynames)
}
\arguments{
\item{amat}{the \code{pcalg}-compatible adjacency matrix of the PAG}

\item{ynames}{names of the variables in set Y}
}
\value{
names of all possible descendants of Y
}
\description{
Gets all possible descendants of Y in the PAG represented by the adjacency matrix amat
}
