% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAGUtils.R
\name{getRegion}
\alias{getRegion}
\title{Gets the region of a set of nodes A}
\usage{
getRegion(amat_v, amat, nodesA)
}
\arguments{
\item{amat_v}{TODO}

\item{amat}{the \code{pcalg}-compatible adjacency matrix of the PAG}

\item{nodesA}{name of the variables in the set A}
}
\value{
the region of A
}
\description{
Gets the region of a set of nodes A
}
