% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAGUtils.R
\name{getVisibleNodesFromX}
\alias{getVisibleNodesFromX}
\title{Gets the nodes that are connected to X by a visible arrow out of X
i.e., returns \{v : x -> v is visible \}}
\usage{
getVisibleNodesFromX(amat, xname)
}
\arguments{
\item{amat}{the \code{pcalg}-compatible adjacency matrix of the PAG}

\item{xname}{name of the variable X}
}
\value{
names of the nodes that are connected to X by a visible arrow out of X
}
\description{
Gets the nodes that are connected to X by a visible arrow out of X
i.e., returns \{v : x -> v is visible \}
}
