% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAGCalculus.R
\name{getXLowerManipulatedPAG}
\alias{getXLowerManipulatedPAG}
\title{Applies the XLower manipulation to the PAG, where all edges that are outgoing from X are removed.}
\usage{
getXLowerManipulatedPAG(amat, xnames)
}
\arguments{
\item{amat}{the \code{pcalg}-compatible adjacency matrix for the PAG.}

\item{xnames}{the names for the variables in the set X}
}
\value{
the \code{pcalg}-compatible adjacency matrix of the
XLower manipulated PAG, \eqn{P_{\underline{x}}}
}
\description{
Applies the XLower manipulation to the PAG, where all edges that are outgoing from X are removed.
}
