% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAGCalculus.R
\name{getXUpperManipulatedPAG}
\alias{getXUpperManipulatedPAG}
\title{Applies the XUpper manipulation to the PAG, where all edges that are incoming to X are removed.}
\usage{
getXUpperManipulatedPAG(amat, xnames)
}
\arguments{
\item{amat}{the \code{pcalg}-compatible adjacency matrix for the PAG.}

\item{xnames}{the names for the variables in the set X}
}
\value{
the \code{pcalg}-compatible adjacency matrix of the
XUpper manipulated PAG, \eqn{P_{\overline{X}}}
}
\description{
Applies the XUpper manipulation to the PAG, where all edges that are incoming to X are removed.
}
