% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDP.R
\name{IDP}
\alias{IDP}
\title{Algorithm for identifying marginal causal effects from PAGs}
\usage{
IDP(amat, x, y, verbose = FALSE)
}
\arguments{
\item{amat}{the pcalg adjacency matrix for the PAG.}

\item{x}{the names for the variables in the treatment set X}

\item{y}{the names for the variables in the outcome set Y}

\item{verbose}{a Boolean indicating whether log messages should be printed out.}
}
\value{
a named list where
  'id'    - a boolean indicating whether P(y|do(x)) is identifiable or not;
  'Qexpr' - list with every required operation in the identification formula as a string;
  'Qop'   - list with every required operation in the identification formula;
  'query' - a string representing P(y|do(x)). It can be used to access the high-level
   identification expression for the query (e.g.,. Qexpr[[query]] and Qop[[query]])
}
\description{
Decides whether the interventional distribution P(y|do(x)) is
identifiable or not given a PAG and the observational distribution P(V).
}
\seealso{
\code{\link{CIDP}}
}
