% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAGUtils.R
\name{inducedPag}
\alias{inducedPag}
\title{Gets the PAG induced over the variables in a set C}
\usage{
inducedPag(amat, cnames, retpag = TRUE)
}
\arguments{
\item{amat}{the \code{pcalg}-compatible adjacency matrix of the PAG}

\item{cnames}{names of the variables in set C}

\item{retpag}{Boolean indicating whether a dagitty PAG should also be computed.}
}
\value{
a \code{pcalg}-compatible adjacency matrix for the induced PAG
}
\description{
Gets the PAG induced over the variables in a set C
}
