% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAGUtils.R
\name{isMSeparated}
\alias{isMSeparated}
\title{Determines whether X is m-separated of Y given S in the PAG represented by its adjacency matrix}
\usage{
isMSeparated(amat, xnames, ynames, snames, verbose = FALSE)
}
\arguments{
\item{amat}{the \code{pcalg}-compatible adjacency matrix of the PAG}

\item{xnames}{the names for the variables in the set X}

\item{ynames}{the names for the variables in the set Y}

\item{snames}{the names for the variables in the set S}

\item{verbose}{a Boolean indicating whether log messages should be printed out.}
}
\value{
Boolean indicating whether X is m-separated of Y given S in the PAG
}
\description{
Determines whether X is m-separated of Y given S in the PAG represented by its adjacency matrix
}
