% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAGCalculus.R
\name{Rule2}
\alias{Rule2}
\title{Verifies whether do-calculus' Rule 2 applies., i.e., if \cr
X and Y are m-separated by \eqn{W \cup Z} in the PAG \eqn{P_{\overline{W}, \underline{X}}}}
\usage{
Rule2(amat, x, y, z, w)
}
\arguments{
\item{amat}{the \code{pcalg}-compatible adjacency matrix for the PAG.}

\item{x}{the names for the variables in the set X}

\item{y}{the names for the variables in the set Y}

\item{z}{the names for the variables in the set Z}

\item{w}{the names for the variables in the set W}
}
\value{
a boolean indicating whether Rule 2 applies
}
\description{
Verifies whether do-calculus' Rule 2 applies., i.e., if \cr
X and Y are m-separated by \eqn{W \cup Z} in the PAG \eqn{P_{\overline{W}, \underline{X}}}
}
