% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAGCalculus.R
\name{Rule3}
\alias{Rule3}
\title{Verifies whether do-calculus' Rule 3 applies., i.e., if X and Y are m-separated by
\eqn{W \cup Z in the PAG P_{\overline{W}, \overline{X(Z)}}, where X(Z) := X \PossAn(Z) in P_{v\setminus w}}}
\usage{
Rule3(amat, x, y, z, w)
}
\arguments{
\item{amat}{the \code{pcalg}-compatible adjacency matrix for the PAG.}

\item{x}{the names for the variables in the set X}

\item{y}{the names for the variables in the set Y}

\item{z}{the names for the variables in the set Z}

\item{w}{the names for the variables in the set W}
}
\value{
a boolean indicating whether Rule 3 applies
}
\description{
Verifies whether do-calculus' Rule 3 applies., i.e., if X and Y are m-separated by
\eqn{W \cup Z in the PAG P_{\overline{W}, \overline{X(Z)}}, where X(Z) := X \PossAn(Z) in P_{v\setminus w}}
}
