% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAGUtils.R
\name{visibleEdgeByNames}
\alias{visibleEdgeByNames}
\title{Determines if the edge v1 -> V2 is visible in the PAG}
\usage{
visibleEdgeByNames(amat, v1name, v2name)
}
\arguments{
\item{amat}{the \code{pcalg}-compatible adjacency matrix of the PAG}

\item{v1name}{name of the variable V1}

\item{v2name}{name of the variable V2}
}
\value{
Boolean indicating if v1 -> V2 is a visible edge
}
\description{
Determines if the edge v1 -> V2 is visible in the PAG
}
